pyproject.toml
[project]
name = "genai-session2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jinja2>=3.1.6",
    "openai>=2.6.1",
    "pandas>=2.3.3",
    "pydantic>=2.12.3",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
